Use memoryview for password gen.
I actually don't know if this saves RAM.  On the one hand, the mem view
lets us keep a list of the indexes rather than the substrings, which I
think should be smaller (unless Python is doing Smart Things under the
hood?)  On the other hand, a list of two-tuples of int might be larger
than a list of substrings!  I didn't check.  You could convert the list
of indicies to an array...

Anyhoo, this is just a bit of code golf on a sunny morning.  I should go
water the garden.